+++ CMSIS/NN/Source/ConvolutionFunctions/arm_nn_mat_mult_s8.c
@@ -21,8 +21,8 @@
  * Title:        arm_nn_mat_mult_s8.c
  * Description:  General Matrix-multiplication function
  *
- * $Date:        March 6, 2020
- * $Revision:    V.2.0.2
+ * $Date:        May 29, 2020
+ * $Revision:    V.2.0.3
  *
  * Target Processor:  Cortex-M cores
  * -------------------------------------------------------------------- */
@@ -73,7 +73,7 @@ q7_t *arm_nn_mat_mult_s8(const q7_t *input_row,
 
             for (int i_row_loop = 0; i_row_loop < row_loop_cnt; i_row_loop++)
             {
-                mve_pred16_t p = vctp16q(row_len_tmp);
+                mve_pred16_t p = vctp16q((uint32_t)row_len_tmp);
                 const int16x8_t offset = vdupq_m_n_s16(vuninitializedq_s16(), col_offset, p);
                 row_len_tmp -= 8;
 
@@ -129,7 +129,7 @@ q7_t *arm_nn_mat_mult_s8(const q7_t *input_row,
 
                 for (int i_row_loop = 0; i_row_loop < row_loop_cnt; i_row_loop++)
                 {
-                    const mve_pred16_t p = vctp16q(row_len_tmp);
+                    const mve_pred16_t p = vctp16q((uint32_t)row_len_tmp);
                     const int16x8_t offset = vdupq_m_n_s16(vuninitializedq_s16(), col_offset, p);
                     row_len_tmp -= 8;
